Split attribute completion module into attribute, derive and lint modules
